Move components to the sepatrate files
