+++ src/ggrc-client/js/components/assessment_templates/templates/attribute_field.mustache
