+++ src/ggrc-client/js/controllers/modals/modals_controller.js
@@ -8,7 +8,8 @@ import '../../components/issue-tracker/issue-tracker-switcher';
 import '../../components/access_control_list/access-control-list-roles-helper';
 import '../../components/assessment/assessment-people';
 import '../../components/assessment/assessment-object-type-dropdown';
-import '../../components/assessment_attributes';
+import '../../components/assessment_templates/assessment-template-attributes';
+import '../../components/assessment_templates/template-field';
 import '../../components/assessment_templates/add-template-field';
 import '../../components/textarea-array/textarea-array';
 import '../../components/object-list-item/object-list-item-updater';
